+++ phoneClients/android/app/src/main/AndroidManifest.xml
@@ -1,17 +1,16 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools">
-    <!-- package attribute removed, namespace is defined in build.gradle -->
-    <!-- Required Permissions -->
+
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
-    <uses-permission android:name="android.permission.ACCESS_COARSE_LOCATION" /> <!-- Required for background location access on Android 10 (API 29) and above -->
-    <uses-permission android:name="android.permission.ACCESS_BACKGROUND_LOCATION" /> <!-- Required for Foreground Service -->
+    <uses-permission android:name="android.permission.ACCESS_COARSE_LOCATION" />
+    <uses-permission android:name="android.permission.ACCESS_BACKGROUND_LOCATION" />
     <uses-permission android:name="android.permission.FOREGROUND_SERVICE" />
-    <uses-permission android:name="android.permission.FOREGROUND_SERVICE_LOCATION" /> <!-- Required for Notification on Android 13+ -->
-    <uses-permission android:name="android.permission.POST_NOTIFICATIONS" /> <!-- Required for Wake Lock -->
-    <uses-permission android:name="android.permission.WAKE_LOCK" /> <!-- Required for Network State -->
-    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" /> <!-- Add clear text traffic permission for testing -->
+    <uses-permission android:name="android.permission.FOREGROUND_SERVICE_LOCATION" />
+    <uses-permission android:name="android.permission.POST_NOTIFICATIONS" />
+    <uses-permission android:name="android.permission.WAKE_LOCK" />
+    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
     <uses-permission android:name="android.permission.CHANGE_NETWORK_STATE" />
 
     <application
@@ -24,21 +23,18 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme.Splash"
         android:usesCleartextTraffic="true"
-        tools:targetApi="33">
+        tools:targetApi="35">
         <activity
             android:name=".MainActivity"
             android:exported="true"
-            android:label="@string/app_name">
+            android:screenOrientation="portrait">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <!-- Declare the Foreground Service -->
         <service
             android:name=".service.TrackingService"
             android:foregroundServiceType="location" />
     </application>
-
 </manifest>
