--- phoneClients/android/app/src/main/java/com/websmithing/gpstracker2/data/repository/ForegroundLocationRepository.kt
@@ -1,16 +1,0 @@
-package com.websmithing.gpstracker2.data.repository
-
-import android.location.Location
-import kotlinx.coroutines.flow.Flow
-
-/**
- * Provides a most recent location data in foreground mode and is not related to the logic of
- * sending tracking data over the network
- */
-interface ForegroundLocationRepository {
-    val currentLocation: Flow<Location?>
-
-    fun start()
-
-    fun stop()
-}
