+++ phoneClients/android/app/src/main/java/com/websmithing/gpstracker2/di/SettingsRepositoryEntryPoint.kt
@@ -1,6 +1,6 @@
 package com.websmithing.gpstracker2.di
 
-import com.websmithing.gpstracker2.data.repository.SettingsRepository
+import com.websmithing.gpstracker2.repository.settings.SettingsRepository
 import dagger.hilt.EntryPoint
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
