--- phoneClients/android/app/src/main/java/com/websmithing/gpstracker2/di/UtilModule.kt
@@ -1,2 +1,0 @@
-// This file is intentionally left empty to resolve build conflicts.
-// Bindings are now in BindsUtilModule.kt and ProvidesUtilModule.kt.
