+++ phoneClients/android/app/src/main/java/com/websmithing/gpstracker2/repository/location/ForegroundLocationRepository.kt
@@ -1,0 +1,10 @@
+package com.websmithing.gpstracker2.repository.location
+
+import android.location.Location
+import kotlinx.coroutines.flow.Flow
+
+interface ForegroundLocationRepository {
+    val currentLocation: Flow<Location?>
+    fun start()
+    fun stop()
+}
