+++ phoneClients/android/app/src/main/java/com/websmithing/gpstracker2/service/TrackingService.kt
@@ -1,4 +1,3 @@
-// # android/app/src/main/java/com/websmithing/gpstracker2/service/TrackingService.kt
 package com.websmithing.gpstracker2.service
 
 import android.annotation.SuppressLint
@@ -23,9 +22,9 @@ import com.google.android.gms.location.LocationRequest
 import com.google.android.gms.location.LocationResult
 import com.google.android.gms.location.Priority
 import com.websmithing.gpstracker2.R
-import com.websmithing.gpstracker2.data.repository.LocationRepository
-import com.websmithing.gpstracker2.data.repository.SettingsRepository
 import com.websmithing.gpstracker2.di.SettingsRepositoryEntryPoint
+import com.websmithing.gpstracker2.repository.location.LocationRepository
+import com.websmithing.gpstracker2.repository.settings.SettingsRepository
 import com.websmithing.gpstracker2.util.LocaleHelper
 import dagger.hilt.android.AndroidEntryPoint
 import dagger.hilt.android.EntryPointAccessors
@@ -39,89 +38,32 @@ import java.util.concurrent.Executors
 import java.util.concurrent.TimeUnit
 import javax.inject.Inject
 
-/**
- * Foreground service responsible for location tracking.
- *
- * This service handles:
- * - Starting and stopping location updates via FusedLocationProviderClient
- * - Processing location data in a background thread
- * - Uploading location data to the remote server
- * - Managing wake locks to ensure tracking continues even when the device is in doze mode
- * - Displaying a persistent notification to inform the user of active tracking
- * - Maintaining the service across app termination and device reboots
- *
- * The service is integrated with Hilt for dependency injection and uses a combination
- * of coroutines (for repository operations) and a single-thread executor for background tasks.
- */
 @AndroidEntryPoint
 class TrackingService : Service() {
 
-    /**
-     * Location provider client for requesting location updates
-     */
     @Inject
     lateinit var fusedLocationProviderClient: FusedLocationProviderClient
 
-    /**
-     * Repository for managing location data operations
-     */
     @Inject
     lateinit var locationRepository: LocationRepository
 
-    /**
-     * Repository for managing app settings
-     */
     @Inject
     lateinit var settingsRepository: SettingsRepository
 
-    /**
-     * HTTP client for network operations
-     */
     @Inject
     lateinit var okHttpClient: OkHttpClient
 
-    /**
-     * Callback for receiving location updates
-     */
     private var locationCallback: LocationCallback? = null
 
-    /**
-     * Executor for running location processing tasks in the background
-     */
     private var backgroundExecutor: ExecutorService? = null
 
-    /**
-     * Wake lock to keep CPU running during tracking
-     */
     private var wakeLock: PowerManager.WakeLock? = null
 
-    /**
-     * Constants used by the service
-     */
     companion object {
-        /**
-         * Intent action to start the service
-         */
         const val ACTION_START_SERVICE = "ACTION_START_SERVICE"
-
-        /**
-         * Intent action to stop the service
-         */
         const val ACTION_STOP_SERVICE = "ACTION_STOP_SERVICE"
-
-        /**
-         * ID for the notification channel
-         */
         private const val NOTIFICATION_CHANNEL_ID = "tracking_channel"
-
-        /**
-         * Name for the notification channel
-         */
         private const val NOTIFICATION_CHANNEL_NAME = "GPS Tracking"
-
-        /**
-         * ID for the service notification
-         */
         private const val NOTIFICATION_ID = 1
     }
 
@@ -135,11 +77,6 @@ class TrackingService : Service() {
         super.attachBaseContext(newCtx)
     }
 
-    /**
-     * Called when the service is first created.
-     *
-     * Initializes the notification channel, wake lock, and background executor.
-     */
     override fun onCreate() {
         super.onCreate()
         Timber.d("TrackingService onCreate")
@@ -185,9 +122,6 @@ class TrackingService : Service() {
         }.start()
     }
 
-    /**
-     * Creates a partial wake lock to keep the CPU running during tracking
-     */
     private fun createWakeLock() {
         val powerManager = getSystemService(Context.POWER_SERVICE) as PowerManager
         wakeLock = powerManager.newWakeLock(
@@ -199,16 +133,6 @@ class TrackingService : Service() {
         Timber.d("Wake lock created")
     }
 
-    /**
-     * Called every time an intent is sent to the service.
-     *
-     * Handles service start/stop requests and manages the foreground state.
-     *
-     * @param intent The intent sent to the service
-     * @param flags Additional data about this start request
-     * @param startId A unique integer representing this specific request to start
-     * @return [START_STICKY] to indicate that the service should be restarted if killed
-     */
     override fun onStartCommand(intent: Intent?, flags: Int, startId: Int): Int {
         Timber.d("TrackingService onStartCommand: ${intent?.action}")
         when (intent?.action) {
@@ -235,13 +159,6 @@ class TrackingService : Service() {
         return START_STICKY
     }
 
-    /**
-     * Called when the user swipes the app away from recent apps.
-     *
-     * Schedules the service to restart after a short delay to ensure continuous tracking.
-     *
-     * @param rootIntent The original intent that was used to launch the task that is being removed
-     */
     override fun onTaskRemoved(rootIntent: Intent?) {
         super.onTaskRemoved(rootIntent)
         Timber.d("TrackingService onTaskRemoved - application swiped away from recent apps")
@@ -264,11 +181,6 @@ class TrackingService : Service() {
         Timber.d("TrackingService scheduled for restart in 5 seconds")
     }
 
-    /**
-     * Called when the service is being destroyed.
-     *
-     * Cleans up resources, stops location updates, and releases the wake lock.
-     */
     override fun onDestroy() {
         super.onDestroy()
         Timber.d("TrackingService onDestroy")
@@ -288,21 +200,10 @@ class TrackingService : Service() {
         wakeLock = null
     }
 
-    /**
-     * Not used in this service implementation.
-     *
-     * @return Always returns null as this is not a bound service
-     */
     override fun onBind(intent: Intent?): IBinder? {
         return null
     }
 
-    /**
-     * Starts location update requests.
-     *
-     * Configures location request parameters based on settings,
-     * sets up location callbacks, and acquires a wake lock.
-     */
     @SuppressLint("MissingPermission")
     private fun startLocationUpdates() {
         Timber.d("Starting location updates...")
@@ -371,15 +272,6 @@ class TrackingService : Service() {
         }
     }
 
-    /**
-     * Processes a new location update.
-     *
-     * Submits the location processing task to the background executor to avoid
-     * blocking the main thread. The background task handles location state updates
-     * and uploading to the server with retry logic.
-     *
-     * @param currentLocation The new location from FusedLocationProviderClient
-     */
     private fun handleNewLocation(currentLocation: Location) {
         Timber.d("handleNewLocation: Received location ${currentLocation.latitude}, ${currentLocation.longitude}")
 
@@ -487,12 +379,6 @@ class TrackingService : Service() {
         } ?: Timber.e("handleNewLocation: Background executor is null, cannot submit task.")
     }
 
-    /**
-     * Stops location updates and releases resources.
-     *
-     * Removes the location callback, releases the wake lock,
-     * and cleans up associated resources.
-     */
     private fun stopLocationUpdates() {
         Timber.d("stopLocationUpdates called.")
 
@@ -524,11 +410,6 @@ class TrackingService : Service() {
         } ?: Timber.d("stopLocationUpdates called but locationCallback was already null.")
     }
 
-    /**
-     * Creates the notification channel for Android O and above.
-     *
-     * This is required for displaying the foreground service notification.
-     */
     private fun createNotificationChannel() {
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
             val channel = NotificationChannel(
@@ -542,14 +423,6 @@ class TrackingService : Service() {
         }
     }
 
-    /**
-     * Creates the notification for the foreground service.
-     *
-     * This notification is displayed while the service is running
-     * to inform the user about the active tracking.
-     *
-     * @return The notification to display
-     */
     private fun createNotification(): Notification {
         return NotificationCompat.Builder(this, NOTIFICATION_CHANNEL_ID)
             .setContentTitle(getString(R.string.app_name))
