+++ phoneClients/android/app/src/main/java/com/websmithing/gpstracker2/ui/components/LabeledBox.kt
@@ -6,7 +6,9 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.websmithing.gpstracker2.ui.theme.WaliotTheme
 
 @Composable
 fun LabeledBox(
@@ -25,5 +27,15 @@ fun LabeledBox(
             modifier = modifier
         )
         content()
+    }
+}
+
+@Preview
+@Composable
+private fun LabeledBoxPreview() {
+    WaliotTheme {
+        LabeledBox(label = "Label") {
+            Text("Content")
+        }
     }
 }
