+++ phoneClients/android/app/src/main/java/com/websmithing/gpstracker2/util/LocaleHelper.kt
@@ -3,15 +3,11 @@ package com.websmithing.gpstracker2.util
 import android.content.Context
 import android.content.res.Configuration
 import android.content.res.Resources
-import com.websmithing.gpstracker2.data.repository.SettingsRepository
+import com.websmithing.gpstracker2.repository.settings.SettingsRepository
 import java.util.Locale
 
 object LocaleHelper {
 
-    /**
-     * Attaches the saved locale to the base context of an Activity or Service.
-     * This must be called in `attachBaseContext`.
-     */
     fun onAttach(context: Context, settingsRepository: SettingsRepository): Context {
         val currentLanguage = settingsRepository.getCurrentLanguage()
         return setAppLocale(context, currentLanguage)
