+++ phoneClients/android/app/src/main/res/values/dimens.xml
@@ -1,3 +1,2 @@
-<resources>
-    <!-- Default screen margins, per the Android Design guidelines. -->
-</resources>
+<?xml version="1.0" encoding="utf-8"?>
+<resources></resources>
