+++ phoneClients/android/app/src/main/res/values/themes.xml
@@ -1,8 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-
     <style name="AppTheme" parent="@style/Theme.AppCompat.DayNight.NoActionBar" />
-
     <style name="AppTheme.Splash" parent="Theme.SplashScreen">
         <item name="windowSplashScreenAnimatedIcon">@drawable/empty</item>
         <item name="windowSplashScreenIconBackgroundColor">@null</item>
